Convert DELETE '/todos/:id' and POST '/users' routes to async/await

+++ server/server.js
@@ -61,24 +61,21 @@ app.get('/todos/:id', authenticate, (req, res) => {
   });
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
   const id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send();
-  } else {
-    Todo.findOneAndDelete({
-      _id: id,
-      _creator: req.user._id
-    }).then((todo) => {
-      if (!todo) {
-        return res.status(404).send();
-      }
+  }
 
-      return res.send({todo});
-    }).catch((e) => {
-      return res.status(400).send(e);
-    });
+  try {
+    const todo = await Todo.findOneAndDelete({_id: id, _creator: req.user._id});
+
+    if (!todo) return res.status(404).send();
+
+    return res.send({todo});
+  } catch (e) {
+    return res.status(400).send(e);
   }
 });
 
@@ -119,43 +116,42 @@ app.patch('/todos/:id', authenticate, (req, res) => {
   });
 });
 
-app.post('/users', (req, res) => {
-  const body = _.pick(req.body, ['email', 'password']);
-  const user = new User(body);
+app.post('/users', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    const user = new User(body);
+    await user.save();
+    const token = await user.generateAuthToken();
 
-  user.save().then(() => {
-    return user.generateAuthToken();
-  }).then((token) => {
-    // console.log(token);
     res.header('x-auth', token).send(user);
-  }, (e) => {
+  } catch (e) {
     res.status(400).send(e);
-  })
+  }
 });
 
 app.get('/users/me', authenticate, (req, res) => {
   res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-  const body = _.pick(req.body, ['email', 'password']);
+app.post('/users/login', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    const user = await User.findByCredentials(body.email, body.password);
+    const token = await user.generateAuthToken();
 
-  User.findByCredentials(body.email, body.password).then((user) => {
-    return user.generateAuthToken().then(((token) => {
-      res.header('x-auth', token)
-        .send(user);
-    }));
-  }).catch((e) => {
+    res.header('x-auth', token).send(user);
+  } catch (e) {
     res.status(400).send(e);
-  });
+  }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-  req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+  try {
+    await req.user.removeToken(req.token);
     res.status(200).send();
-  }, () => {
+  } catch (e) {
     res.status(400).send();
-  });
+  }
 });
 
 app.listen(port, () => {
